[DDW-521] Finish api integration with json stringify rendering
