+++ source/renderer/app/api/transactions/types.js
@@ -3,6 +3,7 @@ import BigNumber from 'bignumber.js';
 import { WalletTransaction } from '../../domains/WalletTransaction';
 import { WalletUnits } from '../../domains/Wallet';
 import type { DelegationAction } from '../../types/stakingTypes';
+import type { TransactionMetadata } from '../../types/TransactionMetadata';
 
 export type TransactionAmount = {
   quantity: number,
@@ -50,6 +51,7 @@ export type Transaction = {
   outputs: Array<TransactionOutputs>,
   withdrawals: Array<TransactionWithdrawals>,
   status: TransactionState,
+  metadata?: TransactionMetadata,
 };
 
 export type Transactions = Array<Transaction>;
