+++ source/renderer/app/components/wallet/transactions/WalletTransactionsList.js
@@ -83,6 +83,7 @@ export default class WalletTransactionsList extends Component<Props> {
   };
 
   expandedTransactionIds: Map<string, WalletTransaction> = new Map();
+  transactionsShowingMetadata: Map<string, WalletTransaction> = new Map();
   virtualList: ?VirtualTransactionList;
   simpleList: ?SimpleTransactionList;
   loadingSpinner: ?LoadingSpinner;
@@ -136,6 +137,9 @@ export default class WalletTransactionsList extends Component<Props> {
   isTxExpanded = (tx: WalletTransaction) =>
     this.expandedTransactionIds.has(tx.id);
 
+  isTxShowingMetadata = (tx: WalletTransaction) =>
+    this.transactionsShowingMetadata.has(tx.id);
+
   toggleTransactionExpandedState = (tx: WalletTransaction) => {
     const isExpanded = this.isTxExpanded(tx);
     if (isExpanded) {
@@ -155,6 +159,7 @@ export default class WalletTransactionsList extends Component<Props> {
    * @param tx
    */
   onShowMetadata = (tx: WalletTransaction) => {
+    this.transactionsShowingMetadata.set(tx.id, tx);
     if (this.virtualList) {
       this.virtualList.updateTxRowHeight(tx, true, true);
     } else if (this.simpleList) {
@@ -198,6 +203,7 @@ export default class WalletTransactionsList extends Component<Props> {
           deletePendingTransaction={deletePendingTransaction}
           formattedWalletAmount={formattedWalletAmount}
           isExpanded={this.isTxExpanded(tx)}
+          isShowingMetadata={this.isTxShowingMetadata(tx)}
           isLastInList={isLastInGroup}
           isRestoreActive={isRestoreActive}
           onDetailsToggled={() => this.toggleTransactionExpandedState(tx)}
