+++ source/renderer/app/components/wallet/transactions/metadata/TransactionMetadataView.scss
@@ -8,51 +8,58 @@
   }
 }
 
-.list {
-  &::before {
-    content: '[';
-    display: inline;
-  }
-  &::after {
-    content: ']';
-    display: inline;
-  }
-  li::after {
-    content: ',';
-    display: inline;
-  }
-  li:last-child::after {
-    content: '';
-  }
-  li {
-    padding-left: 10px;
-  }
-  li > * {
-    display: inline-block;
-  }
-}
+/**
+ * NOTE: These styles are currently not used because we simply
+ * JSON.stringify the metadata for transactions. This can be
+ * used as the basis for a more sophisticated implementation
+ * later on:
+ */
 
-.map {
-  vertical-align: top;
-  &::before {
-    content: '{';
-    display: inline;
-  }
-  &::after {
-    content: '}';
-    display: inline;
-  }
-  li::after {
-    content: ',';
-    display: inline;
-  }
-  li:last-child::after {
-    content: '';
-  }
-  li {
-    padding-left: 10px;
-  }
-  li > * {
-    display: inline-block;
-  }
-}
+//.list {
+//  &::before {
+//    content: '[';
+//    display: inline;
+//  }
+//  &::after {
+//    content: ']';
+//    display: inline;
+//  }
+//  li::after {
+//    content: ',';
+//    display: inline;
+//  }
+//  li:last-child::after {
+//    content: '';
+//  }
+//  li {
+//    padding-left: 10px;
+//  }
+//  li > * {
+//    display: inline-block;
+//  }
+//}
+//
+//.map {
+//  vertical-align: top;
+//  &::before {
+//    content: '{';
+//    display: inline;
+//  }
+//  &::after {
+//    content: '}';
+//    display: inline;
+//  }
+//  li::after {
+//    content: ',';
+//    display: inline;
+//  }
+//  li:last-child::after {
+//    content: '';
+//  }
+//  li {
+//    padding-left: 10px;
+//  }
+//  li > * {
+//    display: inline-block;
+//  }
+//}
